requirements.txt
opencv-python
mss
google-generativeai
pynput
pygetwindow
python-dotenv
numpy
customtkinter
packaging
